refactor: accommodation repository migrated to mongodb
